+++ librina/src/rina-syscalls.h
@@ -56,7 +56,7 @@ namespace rina {
          *         otherwise
          */
         int syscallCreateIPCProcess(const ApplicationProcessNamingInformation & ipcProcessName,
-                                    unsigned int                                ipcProcessId,
+                                    unsigned short                              ipcProcessId,
                                     const std::string &                         difType);
 
         /**
