+++ rinad/src/ipcmanager/src/main/java/rina/ipcmanager/impl/console/UnregisterIPCProcessFromNMinusOneDIF.java
@@ -16,7 +16,7 @@ public class UnregisterIPCProcessFromNMinusOneDIF extends ConsoleCommand{
 	private static final String USAGE = "unregn1dif <ipcp_id> <dif_name>";
 	
 	private CancelCommandTimerTask cancelCommandTask = null;
-	private long ipcProcessId;
+	private int ipcProcessId;
 	private String difName;
 	
 	public UnregisterIPCProcessFromNMinusOneDIF(IPCManager ipcManager, IPCManagerConsole console){
@@ -30,7 +30,7 @@ public class UnregisterIPCProcessFromNMinusOneDIF extends ConsoleCommand{
 		}
 		
 		try{
-			ipcProcessId = Long.parseLong(splittedCommand[1]);
+			ipcProcessId = Integer.parseInt(splittedCommand[1]);
 		} catch(Exception ex) {
 			return "Could not parse ipcProcessId " + splittedCommand[1];
 		}
